requirements.txt
datasets >= 1.2.1
transformers
wikipedia
googlesearch-python